+++ packages/calendar/src/components/ui/date-time-input/DateTextInput.stories.tsx
@@ -1,6 +1,9 @@
 import { Store, withState } from "@dump247/storybook-state";
 import { DateTextInput } from "@stenajs-webui/calendar";
-import { createDateTextInputTheme } from "@stenajs-webui/theme";
+import {
+  createDateTextInputTheme,
+  createStandardTextInputTheme
+} from "@stenajs-webui/theme";
 import { color } from "@storybook/addon-knobs";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
@@ -110,9 +113,11 @@ storiesOf("calendar/Input/DateTextInput", module)
       <DateTextInput
         value={store.state.value}
         onValueChange={value => store.set({ value })}
-        textInputTheme={createDateTextInputTheme({
-          backgroundColor: "lightgoldenrodyellow",
-          backgroundColorInvalidDate: "lightcoral"
+        theme={createDateTextInputTheme({
+          textInput: createStandardTextInputTheme({
+            backgroundColor: "lightgoldenrodyellow",
+            backgroundColorInvalid: "lightcoral"
+          })
         })}
       />
     ))